+++ src/glikoz/__init__.py
@@ -1,2 +1,2 @@
 from .dataframe_handler import DiaguardCSVParser, DataFrameHandler
-from .report_creator import JSONReportCreator, PDFReportCreator
+from .report import JSONReportCreator, PDFReportCreator
